fix(guardian): extract the shared plain-exit-code adapter main() to clear SonarCloud duplication on Kiro's PR
kiro-guardian-adapter.js's main() mirrored windsurf-guardian-adapter.js's
closely enough (44 lines) to trip the new-code duplication gate (14.4%,
required <= 3%) -- both use the same plain exit-code-2-plus-stderr
block contract, unlike Cursor's JSON-envelope variant. Extracted the
shared stdin-read/truncation/exit-code dispatch into
runPlainExitCodeGuardianAdapter() in adapter-stdin-json.js and
migrated both Kiro and Windsurf onto it, removing the duplicate at
the source instead of leaving two parallel copies (the same mistake
already found and fixed once tonight for the hooks-merge side of
this same PR pair).
