+++ scripts/hooks/kiro-guardian-adapter.js
@@ -27,7 +27,7 @@
 'use strict';
 
 const { run } = require('./pre-bash-guardian-validate');
-const { readAdapterStdinJson } = require('../lib/adapter-stdin-json');
+const { runPlainExitCodeGuardianAdapter } = require('../lib/adapter-stdin-json');
 
 const BASH_TOOL_NAMES = new Set(['execute_bash', 'shell', 'execute_cmd']);
 
@@ -50,38 +50,7 @@ function buildGuardianInput(kiroEvent) {
 }
 
 function main() {
-  readAdapterStdinJson(({ ok, truncated, value }) => {
-    if (!ok) {
-      // A parse failure caused by hitting the size cap is not the same as
-      // ordinary malformed input: an attacker can pad a command past the
-      // stdin cap specifically to land here and fail open. Only genuinely
-      // malformed (non-truncated) input gets the fail-open policy the other
-      // adapters use; a truncated payload is unanalyzable and fails closed.
-      if (truncated) {
-        process.stderr.write(
-          'EGC Guardian BLOCKED this command: the event payload exceeded the size ' +
-          'this validator can safely read, so it could not be parsed or validated. ' +
-          'Simplify the command.\n'
-        );
-        process.exit(2);
-      }
-      process.exit(0);
-    }
-
-    const guardianInput = buildGuardianInput(value);
-    if (!guardianInput) {
-      process.exit(0);
-    }
-
-    const result = run(guardianInput);
-    if (result.exitCode === 2) {
-      const reason = result.stderr || 'Blocked by the EGC Guardian.';
-      process.stderr.write(reason.endsWith('\n') ? reason : `${reason}\n`);
-      process.exit(2);
-    }
-
-    process.exit(0);
-  });
+  runPlainExitCodeGuardianAdapter(buildGuardianInput, run);
 }
 
 if (require.main === module) {
