+++ scripts/lib/adapter-stdin-json.js
@@ -35,4 +35,45 @@ function readAdapterStdinJson(onComplete) {
   });
 }
 
-module.exports = { MAX_STDIN, readAdapterStdinJson };
+// Shared main() body for host adapters that block with a plain exit-code-2-
+// plus-stderr contract (Windsurf, Kiro) -- as opposed to Cursor, which also
+// needs a {permission, ...} JSON envelope on stdout and so builds its own
+// main() around readAdapterStdinJson directly instead of this helper.
+// buildGuardianInput: (parsedEvent) => Guardian input object | null.
+// runGuardian: pre-bash-guardian-validate.js's own run().
+function runPlainExitCodeGuardianAdapter(buildGuardianInput, runGuardian) {
+  readAdapterStdinJson(({ ok, truncated, value }) => {
+    if (!ok) {
+      // A parse failure caused by hitting the size cap is not the same as
+      // ordinary malformed input: an attacker can pad a command past the
+      // stdin cap specifically to land here and fail open. Only genuinely
+      // malformed (non-truncated) input gets the fail-open policy the other
+      // adapters use; a truncated payload is unanalyzable and fails closed.
+      if (truncated) {
+        process.stderr.write(
+          'EGC Guardian BLOCKED this command: the event payload exceeded the size ' +
+          'this validator can safely read, so it could not be parsed or validated. ' +
+          'Simplify the command.\n'
+        );
+        process.exit(2);
+      }
+      process.exit(0);
+    }
+
+    const guardianInput = buildGuardianInput(value);
+    if (!guardianInput) {
+      process.exit(0);
+    }
+
+    const result = runGuardian(guardianInput);
+    if (result.exitCode === 2) {
+      const reason = result.stderr || 'Blocked by the EGC Guardian.';
+      process.stderr.write(reason.endsWith('\n') ? reason : `${reason}\n`);
+      process.exit(2);
+    }
+
+    process.exit(0);
+  });
+}
+
+module.exports = { MAX_STDIN, readAdapterStdinJson, runPlainExitCodeGuardianAdapter };
